+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changes in HTMLDOC v1.9.7
 
+- Refactored the PRE rendering code to work around compiler optimization bugs
+  (Issue #349)
 - Fixed a table rowspan + valign bug (Issue #360)